Rename agent files.
